+++ packages/terra-clinical-action-header/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* prop `keepCloseButton` to keep the close button in small viewports
 
 1.6.0 - (December 9, 2017)
 ----------